Split treasury source context mapping
