+++ tests/unit/dpm/construction/test_source_product_context.py
@@ -17,11 +17,13 @@ from src.api.services.construction_liquidity_source_context import (
     source_status_to_method_status,
 )
 from src.api.services.construction_source_product_context import (
-    external_treasury_currency_overlay_context,
     source_product_authority_context_updates,
 )
 from src.api.services.construction_transaction_cost_source_context import (
     transaction_cost_context_from_curve,
+)
+from src.api.services.construction_treasury_source_context import (
+    external_treasury_currency_overlay_context,
 )
 from src.core.construction.models import ConstructionAuthorityContext
 from src.core.construction.vocabulary import ConstructionMethodStatus
